requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.4.2
tensorflow==2.16.1
gdown
scikit-learn
joblib